\name{Garfield1964}
\docType{data}
\alias{Garfield1964}
\title{Dataset for Garfield1964}
\description{
  RcnaData object as returned by the data gathering tool for the seed
  publication \dQuote{The use of citation data in writing the history of
  science} by Garfield, E. and others (1964), with 3 levels of sampling at a
  rate of 5 percent.
  It contains a total of 1426 articles and 1758 relations among these articles.

  The full file name of the Sqlite database that contains the information of
  this dataset can be found at:

  \code{system.file("data/Garfield1964.sqlite", package="citenet")}
}
\usage{data(Garfield1964)}
\format{A \code{RcnaData} object pointing to a SQLite database.}
\references{
  Garfield, E. and others (1964)
  \emph{The use of citation data in writing the history of science}
}
\keyword{datasets}
