\name{Noruzi2005}
\docType{data}
\alias{Noruzi2005}
\title{Dataset for noruzi2005google}
\description{
  RcnaData object as returned by the data gathering tool function for the seed
  publication \dQuote{Google Scholar: The new generation of citation indexes}
  by Noruzi, A. (2005), with 3 levels of sampling at a rate of 10 percent.
  It contains a total of 118 articles and 177 relations among these articles.

  The full file name of the Sqlite database that contains the information of
  this dataset can be found at:

  \code{system.file("data/Noruzi2005.sqlite", package="cna")}
}
\usage{data(Noruzi2005)}
\format{A \code{RcnaData} object pointing to a SQLite database.}
\references{
  Noruzi, A. (2005) 
  \emph{Google Scholar: The new generation of citation indexes}
}
\keyword{datasets}
